requirements.txt
pandas==2.0.3
transformers==4.37.1
ragas==0.0.22
pypdf==4.0.0
python-dotenv==1.0.1